+++ src/app_genesis/routes/__init__.py
@@ -1,0 +1,4 @@
+from .browser import browser
+from .runner import runner
+
+__all__ = ["browser", "runner"]
